refactor: reorganize agent tools under toolsets package
Move docs_mcp_toolset and external tools from agents/tools/ into
agents/toolsets/ and split external tool wiring into its own
build_external_toolset helper.
