--- src/phoenix/server/agents/tools/external_tools/__init__.py
@@ -1,23 +1,0 @@
-from phoenix.server.agents.tools.external_tools.ask_user import ASK_USER_TOOL_DEFINITION
-from phoenix.server.agents.tools.external_tools.bash import BASH_TOOL_DEFINITION
-from phoenix.server.agents.tools.external_tools.clone_prompt_instance import (
-    CLONE_PROMPT_INSTANCE_TOOL_DEFINITION,
-)
-from phoenix.server.agents.tools.external_tools.edit_prompt import EDIT_PROMPT_TOOL_DEFINITION
-from phoenix.server.agents.tools.external_tools.read_prompt import READ_PROMPT_TOOL_DEFINITION
-from phoenix.server.agents.tools.external_tools.set_spans_filter import (
-    SET_SPANS_FILTER_TOOL_DEFINITION,
-)
-from phoenix.server.agents.tools.external_tools.set_time_range import (
-    SET_TIME_RANGE_TOOL_DEFINITION,
-)
-
-__all__ = [
-    "ASK_USER_TOOL_DEFINITION",
-    "BASH_TOOL_DEFINITION",
-    "CLONE_PROMPT_INSTANCE_TOOL_DEFINITION",
-    "EDIT_PROMPT_TOOL_DEFINITION",
-    "READ_PROMPT_TOOL_DEFINITION",
-    "SET_SPANS_FILTER_TOOL_DEFINITION",
-    "SET_TIME_RANGE_TOOL_DEFINITION",
-]
