+++ src/phoenix/server/agents/toolsets/external/tools/bash.py
