+++ src/phoenix/server/app.py
@@ -81,7 +81,7 @@ from phoenix.db.bulk_inserter import BulkInserter
 from phoenix.db.facilitator import Facilitator
 from phoenix.db.helpers import SupportedSQLDialect
 from phoenix.db.insertion.types import AnnotationPrecursor
-from phoenix.server.agents.tools.docs_mcp_toolset import build_docs_mcp_toolset
+from phoenix.server.agents.toolsets.docs_mcp import build_docs_mcp_toolset
 from phoenix.server.api.auth_messages import AUTH_ERROR_MESSAGES, AuthErrorCode
 from phoenix.server.api.context import Context, DataLoaders
 from phoenix.server.api.dataloaders import (
